ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
openai==1.10.0
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-multipart==0.0.6
requests==2.31.0
regex==2023.12.25
